reorganise yaw lock as class
